Dockerfile: use args to better manage UID
The default UID in the container is 1000. When running on a single
machine it is often ok, and matching the UID on the main host. However
when running on a multi user machine, it's preferred to use the UID of
the local user. With this change, we can customize the UID in the
container to allow that. It requires to rebuild the image locally on
each machine, but it is convenient for developers.

To build an image with uses the local user UID:

$ docker build --build-arg UID=$(id -u) -t docker-me

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>